requirements.txt
streamlit
transformers
torch
ftfy
regex
tqdm
pillow
scikit-learn
pandas
numpy
requests
